x/tools/gopls: run go generate through CodeLens
This change adds support for recognizing a //go:generate directive
and offering a CodeLens that will then send a "generate" command to
the server to run "go generate" or "go generate ./...". Because
"go generate" can only be executed per package, there is no need to show
the CodeLens on top of every //go:generate comment. Therefore, only the
top directive will be considered.

The stdout/stderr of the go generate command will be piped to the logger
while stderr will also be sent to the editor as a $/progress notification.
Stderr in this case shows every command that go generate is running through
the "-x" flag so that the user can be aware of the work in progress.

Fixes golang/go#37680

Change-Id: I89a9617521eab20859cb2215db133f34fda856c7